services/data_architecture/pyproject.toml
[tool.poetry]
name = "data_architecture"
version = "0.1.0"
description = "Enhanced AOS Data Architecture Service"
authors = ["Enhanced AOS <aos@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
asyncpg = "^0.29.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.0"
numpy = "^1.26.0"
kafka-python = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
